Move roi tests into a separate file
